indexserver: move sourcegraph api functions into own file
